remove the inheritance from ObjectDetectionBaseModel. The prediction function as a helper will be re-used for FasterRCNN in the next PR.
